+++ src/View/Helper/LibraryHelper.php
@@ -23,7 +23,6 @@
  */
 namespace MeTools\View\Helper;
 
-use Cake\Core\Configure;
 use Cake\View\Helper;
 use MeTools\Core\Plugin;
 
@@ -57,7 +56,7 @@ class LibraryHelper extends Helper {
 	protected function _datetimepicker($input, array $options = []) {
 		$this->Html->js([
 			'/vendor/moment/moment-with-locales.min',
-			'/vendor/bootstrap-datetimepicker/js/bootstrap-datetimepicker.min'
+			'/vendor/bootstrap-datetimepicker/js/bootstrap-datetimepicker.min',
 		], ['block' => 'script_bottom']);
 		
         $this->Html->css('/vendor/bootstrap-datetimepicker/css/bootstrap-datetimepicker.min', ['block' => 'css_bottom']);
@@ -76,7 +75,7 @@ class LibraryHelper extends Helper {
 			'previous' => 'fa fa-arrow-left',
 			'next' => 'fa fa-arrow-right',
 			'today' => 'fa fa-dot-circle-o',
-			'clear' => 'fa fa-trash'
+			'clear' => 'fa fa-trash',
 		], $options);
 		
 		//Sets the current locale
@@ -87,7 +86,8 @@ class LibraryHelper extends Helper {
 	}
 
     /**
-     * Before layout callback. beforeLayout is called before the layout is rendered.
+     * Before layout callback. beforeLayout is called before the layout is 
+     *  rendered.
 	 * @param \Cake\Event\Event $event An Event instance
      * @param string $layoutFile The layout about to be rendered
 	 * @uses MeTools\View\Helper\HtmlHelper::scriptBlock()
@@ -114,8 +114,9 @@ class LibraryHelper extends Helper {
 	 * @return mixed Html code or NULL if is localhost
 	 */
 	public function analytics($id) {
-		if(is_localhost())
+		if(is_localhost()) {
 			return NULL;
+        }
 		
 		return $this->Html->scriptBlock(sprintf('!function(e,a,t,n,c,o,s){e.GoogleAnalyticsObject=c,e[c]=e[c]||function(){(e[c].q=e[c].q||[]).push(arguments)},e[c].l=1*new Date,o=a.createElement(t),s=a.getElementsByTagName(t)[0],o.async=1,o.src=n,s.parentNode.insertBefore(o,s)}(window,document,"script","//www.google-analytics.com/analytics.js","ga"),ga("create","%s","auto"),ga("send","pageview");', $id), ['block' => 'script_bottom']);
 	}
@@ -123,11 +124,13 @@ class LibraryHelper extends Helper {
 	/**
      * Loads all CKEditor scripts.
      * 
-     * To know how to install and configure CKEditor, please refer to the `README.md` file.
+     * To know how to install and configure CKEditor, please refer to the 
+     *  `README.md` file.
 	 * 
 	 * CKEditor must be located into `APP/webroot/ckeditor`.
      * 
-     * To create an input field for CKEditor, you should use the `ckeditor()` method provided by the `FormHelper`.
+     * To create an input field for CKEditor, you should use the `ckeditor()` 
+     *  method provided by the `FormHelper`.
      * @param bool $jquery FALSE if you don't want to use the jQuery adapter
      * @see MeTools\View\Helper\FormHelper::ckeditor()
      * @see http://docs.cksource.com CKEditor documentation
@@ -136,32 +139,38 @@ class LibraryHelper extends Helper {
     public function ckeditor($jquery = TRUE) {
 		$path = WWW_ROOT.DS.'ckeditor'.DS;
 		
-		if(!is_readable($path.'ckeditor.js'))
+		if(!is_readable($path.'ckeditor.js')) {
 			return; 
+        }
 
 		$scripts = ['/ckeditor/ckeditor'];
 
 		//Checks for the jQuery adapter
-		if($jquery && is_readable($path.DS.'adapters'.DS.'jquery.js'))
+		if($jquery && is_readable($path.DS.'adapters'.DS.'jquery.js')) {
 			$scripts[] = '/ckeditor/adapters/jquery';
+        }
 
 		//Checks for `APP/webroot/js/ckeditor_init.php`
-		if(is_readable(WWW_ROOT.'js'.DS.'ckeditor_init.php'))
+		if(is_readable(WWW_ROOT.'js'.DS.'ckeditor_init.php')) {
 			$scripts[] = 'ckeditor_init.php?';
+        }
 		//Checks for `APP/webroot/js/ckeditor_init.js`
-		elseif(is_readable(WWW_ROOT.'js'.DS.'ckeditor_init.js'))
+		elseif(is_readable(WWW_ROOT.'js'.DS.'ckeditor_init.js')) {
 			$scripts[] = 'ckeditor_init';
+        }
 		//Else, uses `APP/plugin/MeTools/webroot/js/ckeditor_init.js`
-		else
+		else {
 			$scripts[] = 'MeTools.ckeditor_init.php?';
-		
+        }
+        
 		$this->Html->js($scripts, ['block' => 'script_bottom']);
     }
 
     /**
      * Adds a datepicker to the `$input` field.
      * 
-     * To create an input field compatible with datepicker, you should use the `datepicker()` method provided by the `FormHelper`.
+     * To create an input field compatible with datepicker, you should use the 
+     *  `datepicker()` method provided by the `FormHelper`.
 	 * 
 	 * Bootstrap Datepicker and Moment.js should be installed via Composer.
      * @param string $input Target field. Default is `.datepicker`
@@ -182,8 +191,8 @@ class LibraryHelper extends Helper {
 	 /**
      * Adds a datetimepicker to the `$input` field.
      * 
-     * To create an input field compatible with datetimepicker, you should use the `datetimepicker()` method provided by the `FormHelper`.
-	 * 
+     * To create an input field compatible with datetimepicker, you should use 
+     *  the `datetimepicker()` method provided by the `FormHelper`.
 	 * Bootstrap Datepicker and Moment.js should be installed via Composer.
      * @param string $input Target field. Default is `.datetimepicker`
      * @param array $options Options for the datetimepicker
@@ -203,32 +212,35 @@ class LibraryHelper extends Helper {
 	 * 
 	 * FancyBox must be installed via Composer.
      * @see http://fancyapps.com/fancybox/#docs FancyBox documentation
-	 * @uses MeTools\View\Helper\AssetHelper::css()
 	 * @uses MeTools\View\Helper\AssetHelper::js()
 	 * @uses MeTools\Core\Plugin::path()
 	 */
 	public function fancybox() {		
-		$this->Asset->css([
+		$this->Html->css([
 			'/vendor/fancybox/jquery.fancybox',
 			'/vendor/fancybox/helpers/jquery.fancybox-buttons',
-			'/vendor/fancybox/helpers/jquery.fancybox-thumbs'
+			'/vendor/fancybox/helpers/jquery.fancybox-thumbs',
 		], ['block' => 'css_bottom']);
 		
 		$this->Asset->js([
 			'/vendor/fancybox/jquery.fancybox.pack',
 			'/vendor/fancybox/helpers/jquery.fancybox-buttons',
-			'/vendor/fancybox/helpers/jquery.fancybox-thumbs'
+			'/vendor/fancybox/helpers/jquery.fancybox-thumbs',
 		], ['block' => 'script_bottom']);
 		
 		//Checks for the init script into `APP/webroot/js/`
-		if(is_readable(WWW_ROOT.'js'.DS.'fancybox_init.js'))
+		if(is_readable(WWW_ROOT.'js'.DS.'fancybox_init.js')) {
 			$script = 'fancybox_init';
-		//Else, checks for the init script into `APP/plugin/MeTools/webroot/fancybox/`
-		elseif(Plugin::path('MeTools', 'webroot'.DS.'fancybox'.DS.'fancybox_init.js', TRUE))
+        }
+		//Else, checks for the init script into 
+        //  `APP/plugin/MeTools/webroot/fancybox/`
+		elseif(Plugin::path('MeTools', 'webroot'.DS.'fancybox'.DS.'fancybox_init.js', TRUE)) {
 			$script = 'MeTools./fancybox/fancybox_init';
-		else
+        }
+		else {
 			return;
-		
+        }
+        
 		$this->Asset->js($script, ['block' => 'script_bottom']);
 	}
 	
@@ -249,7 +261,8 @@ class LibraryHelper extends Helper {
 	/**
      * Through `slugify.js`, it provides the slug of a field. 
      * 
-     * It reads the value of the `$sourceField` field and it sets its slug in the `$targetField`.
+     * It reads the value of the `$sourceField` field and it sets its slug in 
+     *  the `$targetField`.
      * @param string $sourceField Source field
      * @param string $targetField Target field
 	 * @uses MeTools\View\Helper\AssetHelper::js()
@@ -264,7 +277,8 @@ class LibraryHelper extends Helper {
     /**
      * Adds a timepicker to the `$input` field.
      * 
-     * To create an input field compatible with datepicker, you should use the `timepicker()` method provided by the `FormHelper`.
+     * To create an input field compatible with datepicker, you should use the 
+     *  `timepicker()` method provided by the `FormHelper`.
 	 * 
 	 * Bootstrap Datepicker and Moment.js should be installed via Composer.
      * @param string $input Target field. Default is `.timepicker`